Refactored testing

+++ src/main/java/team/awesome/pet/controller/ShelterController.java
@@ -1,6 +1,5 @@
 package team.awesome.pet.controller;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,30 +18,36 @@ import team.awesome.pet.service.ShelterService;
 public class ShelterController {
 
   @Autowired
+  // private ShelterRepository shelterRepo;
   private ShelterService shelterService;
 
   @RequestMapping(method = RequestMethod.GET)
   public Collection<Shelter> getAllShelter() {
     return this.shelterService.getAllShelter();
+    // return this.shelterRepo.findAll();
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
   public Shelter getShelterById(@PathVariable("id") int id) {
     return this.shelterService.getShelterById(id);
+    // return this.shelterRepo.findOne(id);
   }
 
   @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
   public void insertShelter(@RequestBody Shelter shelter) {
     this.shelterService.insertShelter(shelter);
+    // this.shelterRepo.save(shelter);
   }
 
   @RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
   public void updateShelter(@RequestBody Shelter shelter) {
     this.shelterService.updateShelter(shelter);
+    // this.shelterRepo.save(shelter);
   }
 
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-  public void deleteShelterById(@PathVariable("id") BigInteger id) {
+  public void deleteShelterById(@PathVariable("id") int id) {
     this.shelterService.removeShelterById(id);
+    // this.shelterRepo.delete(id);
   }
 }
